Correcting the code according to the suggestions
